{{ cookiecutter.project_slug }}/requirements/local.txt
-r base.txt

django-browser-reload==1.12.1
django-debug-toolbar==4.2.0